chore(travis): fix the build
I don’t know what the issue really is. There is some weird race condition,
where the task copies the file (`pubspec.yml`) and then start `pub get`
which reads that file and the file is not there or is empty.

I have no idea why moving this task into a separate process fixes the
issue but I don’t wanna waste more time on it.